+++ benito-frontend/src/components/Header/HamburgerMenu/constants.ts
@@ -1,0 +1,2 @@
+export const searchIconUrl =
+  "https://cdn2.iconfinder.com/data/icons/font-awesome/1792/search-512.png";
