+++ benito-frontend/src/components/Header/HamburgerMenu/hamburger.scss
