+++ benito-frontend/src/components/Header/HamburgerMenu/styles.scss
@@ -1,0 +1,23 @@
+.qui-search-icon {
+  left: 24px;
+  top: 20px;
+  width: 24px;
+  height: 24px;
+}
+
+.invert-colors {
+  -webkit-filter: invert(1);
+  filter: invert(1);
+}
+
+.qui-mobile-search-container {
+  background-color: #f1f1f1;
+}
+
+.qui-mobile-search {
+  background-color: #f1f1f1;
+  height: 64px;
+  border-radius: 0px;
+  border: none;
+  text-indent: 16px;
+}
